+++ backend/diabetes.sql
@@ -39,14 +39,14 @@ CREATE TABLE IF NOT EXISTS user_health_info (
 -- 糖尿病数据集表结构
 CREATE TABLE IF NOT EXISTS diabetes_data (
     id INT AUTO_INCREMENT PRIMARY KEY,
-    Pregnancies INT NOT NULL COMMENT '怀孕次数',
-    Glucose INT NOT NULL COMMENT '口服葡萄糖耐量实验中血浆葡萄糖浓度（2小时）',
-    BloodPressure INT NOT NULL COMMENT '舒张压（mmHg)',
-    SkinThickness INT NOT NULL COMMENT '肱三头肌皮褶厚度（mm)',
-    Insulin INT NOT NULL COMMENT '2 小时血清胰岛素（μU/ml)',
+    Pregnancies FLOAT NOT NULL COMMENT '怀孕次数',
+    Glucose FLOAT NOT NULL COMMENT '口服葡萄糖耐量实验中血浆葡萄糖浓度（2小时）',
+    BloodPressure FLOAT NOT NULL COMMENT '舒张压（mmHg)',
+    SkinThickness FLOAT NOT NULL COMMENT '肱三头肌皮褶厚度（mm)',
+    Insulin FLOAT NOT NULL COMMENT '2 小时血清胰岛素（μU/ml)',
     BMI FLOAT NOT NULL COMMENT '体重指数（体重kg/(身高m)^2)',
     DiabetesPedigreeFunction FLOAT NOT NULL COMMENT '糖尿病谱系功能',
-    Age INT NOT NULL COMMENT '年龄（岁）',
+    Age FLOAT NOT NULL COMMENT '年龄（岁）',
     Outcome INT NOT NULL COMMENT '是否患病（0或1)'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
